Refactor DTOs and Models for User and Flight Management
- Updated FlightDTO to remove unused fields and improve constructor.
- Changed UserFlightShiftResponseDTO and UserFlightShiftResponseSearchDTO to use Integer for IDs instead of Long.
- Introduced ApplyFlightShiftRequest DTO for shift applications.
- Enhanced Airport model with validation annotations and changed latitude/longitude types to BigDecimal.
- Modified Flight model to use Airport entities instead of String for departure and arrival airports.
- Updated UserFlightShift model to use Integer for ID and improved relationships with User and Flight.
- Adjusted repositories to use Integer for ID types.
- Refactored Team and Unit services to accommodate Integer IDs.
- Removed RoleServiceImpl as it was redundant.
- Improved UserFlightShiftService with new methods for available and assigned shifts.

+++ src/main/java/com/project/quanlycanghangkhong/service/UnitService.java
@@ -7,9 +7,13 @@ import com.project.quanlycanghangkhong.model.Unit;
 
 public interface UnitService {
     List<UnitDTO> getUnitsByTeam(Integer teamId);
-    Unit getUnitById(Long id);
+
+    Unit getUnitById(Integer id);
+
     Unit createUnit(Unit unit);
-    void deleteUnit(Long id);
+
+    void deleteUnit(Integer id);
+
     List<UnitDTO> getAllUnits();
-    
+
 }
